+++ app/controllers/notes_controller.rb
@@ -2,16 +2,17 @@ class NotesController < ApplicationController
   before_action :find_project
   before_action :find_project_level_notes, only: :index
   after_action :verify_authorized, except: :index
-  after_action :verify_policy_scoped, only: :index
 
   def new
+    authorize @project, :new_note?
+
     @note = Note.new(project: @project, user_id:, task_identifier: params[:task_identifier])
-    authorize @note
   end
 
   def create
+    authorize @project, :new_note?
+
     @note = Note.new(project: @project, user_id:, **note_params)
-    authorize @note
 
     if @note.valid?
       @note.save
@@ -74,7 +75,7 @@ class NotesController < ApplicationController
   end
 
   private def find_project_level_notes
-    @notes = policy_scope(Note).includes([:user]).project_level_notes(@project)
+    @notes = Note.project_level_notes(@project).includes([:user, :project])
   end
 
   private def note_params