+++ app/policies/contact_policy.rb
@@ -1,0 +1,29 @@
+class ContactPolicy
+  attr_reader :user, :contact
+
+  def initialize(user, contact)
+    @user = user
+    @contact = contact
+    @project = @contact.project
+  end
+
+  def edit?
+    return false if @project.completed?
+
+    true
+  end
+
+  def update?
+    edit?
+  end
+
+  def destroy?
+    return false if @project.completed?
+
+    true
+  end
+
+  def confirm_destroy?
+    destroy?
+  end
+end