+++ app/policies/note_policy.rb
@@ -3,19 +3,14 @@ class NotePolicy
 
   def initialize(user, note)
     @user = user
-    @record = note
-  end
-
-  def create?
-    true
-  end
-
-  def new?
-    create?
+    @note = note
+    @project = @note.project
   end
 
   def edit?
-    @record.user == @user
+    return false if @project.completed?
+
+    @note.user == @user
   end
 
   def update?
@@ -28,20 +23,5 @@ class NotePolicy
 
   def confirm_destroy?
     edit?
-  end
-
-  class Scope
-    def initialize(user, scope)
-      @user = user
-      @scope = scope
-    end
-
-    def resolve
-      scope.all
-    end
-
-    private
-
-    attr_reader :user, :scope
   end
 end