+++ app/policies/project_policy.rb
@@ -26,12 +26,46 @@ class ProjectPolicy
     create?
   end
 
+  def update?
+    return false if @record.completed?
+
+    project_assigned_to_user?
+  end
+
   def openers?
     true
   end
 
   def change_conversion_date?
-    @record.conversion_date_provisional? == false
+    return false if @record.conversion_date_provisional?
+
+    project_assigned_to_user?
+  end
+
+  def new_note?
+    edit_project_closed?
+  end
+
+  def new_contact?
+    edit_project_closed?
+  end
+
+  def update_assigned_to?
+    edit_project_closed?
+  end
+
+  def update_regional_delivery_officer?
+    edit_project_closed?
+  end
+
+  private def project_assigned_to_user?
+    @record.assigned_to == @user
+  end
+
+  private def edit_project_closed?
+    return false if @record.completed?
+
+    true
   end
 
   class Scope
@@ -50,8 +84,6 @@ class ProjectPolicy
       end
     end
 
-    private
-
-    attr_reader :user, :scope
+    private attr_reader :user, :scope
   end
 end