+++ spec/requests/conversions/involuntary/task_lists_controller_spec.rb
@@ -1,7 +1,7 @@
 require "rails_helper"
 
 RSpec.describe Conversions::Involuntary::TaskListsController do
-  let(:project) { create(:involuntary_conversion_project) }
+  let(:project) { create(:involuntary_conversion_project, assigned_to: user) }
   let(:project_id) { project.id }
   let(:index_path) { conversions_involuntary_project_task_list_path(project_id) }
   let(:task) { build(:involuntary_conversion_task_handover) }